Feature: Tezos message sync
Problem: Users need a solution to pin content on Aleph
automatically when minting NFTs on Tezos.

Solution: Add a sync worker to fetch messages from the Aleph
sync smart contract on Tezos and transform them into valid
Aleph messages.

+++ src/aleph/chains/tx_context.py
@@ -7,5 +7,5 @@ class TxContext:
     tx_hash: str
     height: int
     # Transaction timestamp, in Unix time (number of seconds since epoch).
-    time: int
+    time: float
     publisher: str
